# Lookup based u8 conversion