Hello Jakob,
this is a relatively trivial  pull request adding lookup table based u8->float conversion, that helped me to get openwebrx running  acceptably on an old "thin client" HP fanless PC.  This  creates an array of 256 floats that saves float mult/div operations when geeting data from rtl_sdr dongle. Please merge if you find it useful.  Note that it is my first first pull request on github so I hope I did not screw it up.   

Vy 73,

Piotr, SP9MUF