@jketterl 2021-07-02
Very interesting idea, trade a kilobyte of memory for better CPU performance. I like it. There's a few questions that I need to answer before i can merge this. If you happen to know the answers already, please let me know, otherwise I'll set up a few benchmarks to find out:

Is the new loop in the converter method still vectorizable? If it isn't how does its performance compare to a vectorized version of the old loop?