@ppmkm 2021-07-05
Sorry, but I don't know the answers, I am not able to look onto it at least until 20th of July (vacation), please do the benchmarks if you can. What I know this performs better than original on an old hp thin client originally sold with windows xp embedded, this is a single core 64 bit AMD CPU. I will get the exact specs  when I am back home.Regards,Piotr
-------- Oryginalna wiadomość --------Od: Jakob Ketterl ***@***.***> Data: 02.07.2021  18:55  (GMT+01:00) Do: jketterl/owrx_connector ***@***.***> DW: ppmkm ***@***.***>, Author ***@***.***> Temat: Re: [jketterl/owrx_connector] Lookup based u8 conversion (#12) 
Very interesting idea, trade a kilobyte of memory for better CPU performance. I like it. There's a few questions that I need to answer before i can merge this. If you happen to know the answers already, please let me know, otherwise I'll set up a few benchmarks to find out:
Is the new loop in the converter method still vectorizable? If it isn't how does its performance compare to a vectorized version of the old loop?

—You are receiving this because you authored the thread.Reply to this email directly, view it on GitHub, or unsubscribe.