@jketterl 2021-07-08
OK so I'm afraid, this is going to be complicated. First of all, results:

To answer the first question, the code does no longer vectorize. I suppose the CPU cannot do multiple memory lookup to random locations in parallel, but I'm no expert at that. That means that I had to look at performance.

For the benchmark, I set up a loop that read 1024 bytes from `/dev/urandom` and converts them to floats using either the old mathematical approach vs. the new lookup approach. I ran this in a loop for a million times. If I did my math correctly, that should be somewhere close to a gigabyte of input (depending on how you interpret a gigabyte).

Here's the results on my desktop (Intel Core i7 7700):

old:
```
$ time src/app/throughput_test 
real    0m19,121s
user    0m0,279s
sys     0m18,758s
```

new:
```
$ time src/app/throughput_test 
real    0m19,328s
user    0m0,433s
sys     0m18,848s
```

It's close, but the old implementation is a little little bit faster. I then went to repeat the same tests on a Raspberry Pi (I'm using a 3B+ simply because that's what I had available):

old:
```
$ time src/app/throughput_test  
real    0m33.229s
user    0m2.131s
sys     0m31.094s
```

new:
```
$ time src/app/throughput_test 
real    0m37.484s
user    0m6.337s
sys     0m31.121s
```

Here, too, the old implementation works faster than the new one, this time with a clear margin.

So I'm afraid that looking at these results, I won't be able to merge this in its current form.

This does not mean that I do not believe that your changes are performing better on your machine. I do expect your modification to be very effective on older CPUs that don't have vectorization, or CPUs that have older generation vectorization instruction sets. The challenge here though is to find out where exactly the new implementation performs better, and then build a framework that would switch the underlying implementation based on that information.

The testing code is in this branch: https://github.com/jketterl/owrx_connector/tree/throughput_testing - it is based on this PR, so it's using the new method by default. To switch it to the old implementation, I switched the function body for `Connector::convert(uint8_t*, float*, uint32_t)` back to the one from before the PR.

**I'd be very interested to see other people's results.**