@ppmkm 2021-07-21
Hi again, I am back home and did some of my own tests. I am afraid your results are skewed by the latency of /dev/urandom, this takes up processing power as well. 
On my desktop machine the results without using /dev/urandom seem to favor lookup over computation despite missing vectorization. That being said it may be worth looking if eliminating division from vectorizable code will not bring even bigger improvement.
Would it be possible for you to run the test  with prepared data to remove the /dev/urandom from equation? 
For detailed info please read on.

CPU info: 

vendor_id	: AuthenticAMD
cpu family	: 16
model		: 4
model name	: AMD Phenom(tm) II X3 720 Processor
stepping	: 2
microcode	: 0x10000db
cpu MHz		: 800.000
cache size	: 512 KB
physical id	: 0
siblings	: 3
core id		: 2
cpu cores	: 3
apicid		: 2
initial apicid	: 2
fpu		: yes
fpu_exception	: yes
cpuid level	: 5
wp		: yes
flags		: fpu vme de pse tsc msr pae mce cx8 apic sep mtrr pge mca cmov pat pse36 clflush mmx fxsr sse sse2 ht syscall nx mmxext fxsr_opt pdpe1gb rdtscp lm 3dnowext 3dnow constant_tsc art rep_good nopl nonstop_tsc extd_apicid pni monitor cx16 popcnt lahf_lm cmp_legacy svm extapic cr8_legacy abm sse4a misalignsse 3dnowprefetch osvw ibs skinit wdt hw_pstate retpoline_amd ibp_disable vmmcall npt lbrv svm_lock nrip_save
bogomips	: 5618.99
TLB size	: 1024 4K pages
clflush size	: 64
cache_alignment	: 64
address sizes	: 48 bits physical, 48 bits virtual
power management: ts ttp tm stc 100mhzsteps hwpstate



**original test (lookup based as checked out from throughtput test git branch):** 
`[piotr@step build]$ time time src/app/throughput_test`
`real	0m5.869s`
`user	0m1.053s`
`sys	0m4.815s`
`[piotr@step build]$ time src/app/throughput_test`
`real	0m6.007s`
`user	0m1.143s`
`sys	0m4.862s`

**just reading urandom takes a lot of time**: 

`time dd if=/dev/urandom bs=1000000 count=1000 of=/dev/null`
`1000+0 przeczytanych recordów`
`1000+0 zapisanych recordów`
`skopiowane 1000000000 bajtów (1,0 GB), 4,6102 s, 217 MB/s`
`real	0m4.615s`
`user	0m0.002s`
`sys	0m4.610`

so I precreated a file with random data: 
`dd if=/dev/urandom bs=1000000 count=1000 of=random.dat`
and that instead as an input to the test: 

The results: 
**random dat 1GB, computed:** 
`[piotr@step build]$ time src/app/throughput_test` 

`real	0m2.275s`
`user	0m1.943s`
`sys	0m0.321s`
`[piotr@step build]$ time src/app/throughput_test`

`real	0m2.347s`
`user	0m2.004s`
`sys	0m0.341s`
`[piotr@step build]$ time src/app/throughput_test `

`real	0m2.293s`
`user	0m1.963s`
`sys	0m0.329s`
`[piotr@step build]$ time src/app/throughput_test `

`real	0m2.251s`
`user	0m1.937s`
`sys	0m0.312s`
`[piotr@step build]$ time src/app/throughput_test `

`real	0m2.226s`
`user	0m1.896s`
`sys	0m0.329s`
`[piotr@step build]$ time src/app/throughput_test `

`real	0m2.267s`
`user	0m1.936s`
`sys	0m0.330s`
`[piotr@step build]$ time src/app/throughput_test `

`real	0m2.271s`
`user	0m1.978s`
`sys	0m0.291s`


**random.dat 1GB lookup table**
`[piotr@step build]$ time src/app/throughput_test `

`real	0m1.486s`
`user	0m1.154s`
`sys	0m0.330s`
`[piotr@step build]$ time src/app/throughput_test `

`real	0m1.509s`
`user	0m1.205s`
`sys	0m0.303s`
`[piotr@step build]$ time src/app/throughput_test `

`real	0m1.575s`
`user	0m1.242s`
`sys	0m0.330s`
`[piotr@step build]$ time src/app/throughput_test `

`real	0m1.542s`
`user	0m1.239s`
`sys	0m0.303s`
`[piotr@step build]$ time src/app/throughput_test `

`real	0m1.599s`
`user	0m1.310s`
`sys	0m0.286s`
`[piotr@step build]$ time src/app/throughput_test `

`real	0m1.624s`
`user	0m1.213s`
`sys	0m0.373s`
`[piotr@step build]$ time src/app/throughput_test `

`real	0m1.793s`
`user	0m1.437s`
`sys	0m0.354s`